+++ lua/plugins/luasnip.lua
@@ -1,0 +1,26 @@
+return {
+  "L3MON4D3/LuaSnip",
+  version = "v2.*",
+  build = "make install_jsregexp",
+  dependencies = { "rafamadriz/friendly-snippets" },
+  config = function()
+    local ls = require("luasnip")
+
+    require("luasnip.loaders.from_lua").lazy_load({
+      paths = vim.fn.stdpath("config") .. "/lua/snippets"
+    })
+
+    require("luasnip.loaders.from_vscode").lazy_load()
+
+    -- Jump Keys
+    vim.keymap.set({ "i", "s" }, "<C-k>", function() ls.jump(1) end, { silent = true })
+    vim.keymap.set({ "i", "s" }, "<C-j>", function() ls.jump(-1) end, { silent = true })
+
+    -- Cycle Choice Key
+    vim.keymap.set({ "i", "s" }, "<C-l>", function()
+      if ls.choice_active() then
+        ls.change_choice(1)
+      end
+    end, { silent = true })
+  end,
+}